Stream/sriya-wf-dynamic/pyproject.toml
[project]
name = "sriya-wf-dynamic"
version = "0.1.0"
description = ""
authors = [
    {name = "sriyaraomuthyala",email = "sriyaraomuthyala@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"
typer = "^0.9.3"


[tool.poetry.scripts]
sriya-wf-dynamic = "sriya_wf_dynamic.process:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


